Fix review comments in entries dashboard
