+++ app/views/Project/Tagging/Dashboard/VisualizationShareModal/styles.css
@@ -1,0 +1,9 @@
+.modal {
+    width: 640px;
+
+    .modal-body {
+        display: flex;
+        flex-direction: column;
+        gap: var(--dui-spacing-medium);
+    }
+}
