+++ app/views/Project/Tagging/Dashboard/styles.css
@@ -26,13 +26,3 @@
         }
     }
 }
-
-.modal {
-    width: 640px;
-
-    .modal-body {
-        display: flex;
-        flex-direction: column;
-        gap: var(--dui-spacing-medium);
-    }
-}
